reformat and cleanup;

+++ tests/ABSmartly.Sdk.Tests/Internal/Hashing/Murmur3Tests.cs
@@ -8,21 +8,21 @@ public class Murmur3Tests
 {
     [TestCase("", (uint)0x00000000, (uint)0x00000000)]
     [TestCase(" ", (uint)0x00000000, (uint)0x7ef49b98)]
-    [TestCase("t", (uint)0x00000000, (uint)0xca87df4d)]
-    [TestCase("te", (uint)0x00000000, (uint)0xedb8ee1b)]
+    [TestCase("t", (uint)0x00000000, 0xca87df4d)]
+    [TestCase("te", (uint)0x00000000, 0xedb8ee1b)]
     [TestCase("tes", (uint)0x00000000, (uint)0x0bb90e5a)]
-    [TestCase("test", (uint)0x00000000, (uint)0xba6bd213)]
+    [TestCase("test", (uint)0x00000000, 0xba6bd213)]
     [TestCase("testy", (uint)0x00000000, (uint)0x44af8342)]
-    [TestCase("testy1", (uint)0x00000000, (uint)0x8a1a243a)]
-    [TestCase("testy12", (uint)0x00000000, (uint)0x845461b9)]
+    [TestCase("testy1", (uint)0x00000000, 0x8a1a243a)]
+    [TestCase("testy12", (uint)0x00000000, 0x845461b9)]
     [TestCase("testy123", (uint)0x00000000, (uint)0x47628ac4)]
-    [TestCase("special characters açb↓c", (uint)0x00000000, (uint)0xbe83b140)]
+    [TestCase("special characters açb↓c", (uint)0x00000000, 0xbe83b140)]
     [TestCase("The quick brown fox jumps over the lazy dog", (uint)0x00000000, (uint)0x2e4ff723)]
-    [TestCase("", (uint)0xdeadbeef, (uint)0x0de5c6a9)]
-    [TestCase(" ", (uint)0xdeadbeef, (uint)0x25acce43)]
-    [TestCase("t", (uint)0xdeadbeef, (uint)0x3b15dcf8)]
-    [TestCase("te", (uint)0xdeadbeef, (uint)0xac981332)]
-    [TestCase("tes", (uint)0xdeadbeef, (uint)0xc1c78dda)]
+    [TestCase("", 0xdeadbeef, (uint)0x0de5c6a9)]
+    [TestCase(" ", 0xdeadbeef, (uint)0x25acce43)]
+    [TestCase("t", 0xdeadbeef, (uint)0x3b15dcf8)]
+    [TestCase("te", 0xdeadbeef, 0xac981332)]
+    [TestCase("tes", 0xdeadbeef, 0xc1c78dda)]
     [TestCase("test", 0xdeadbeef, 0xaa22d41a)]
     [TestCase("testy", 0xdeadbeef, 0x84f5f623)]
     [TestCase("testy1", 0xdeadbeef, (uint)0x09ed28e9)]
@@ -41,14 +41,14 @@ public class Murmur3Tests
     [TestCase("testy12", (uint)0x00000001, 0xd92c9f23)]
     [TestCase("testy123", (uint)0x00000001, (uint)0x3bc1712d)]
     [TestCase("special characters açb↓c", (uint)0x00000001, (uint)0x293327b5)]
-    [TestCase("The quick brown fox jumps over the lazy dog", (uint)0x00000001, (uint)0x78e69e27)]    
+    [TestCase("The quick brown fox jumps over the lazy dog", (uint)0x00000001, (uint)0x78e69e27)]
     public void TestDigest(string input, uint seedHex, uint expectedHex)
     {
         var bytes = Encoding.UTF8.GetBytes(input);
         var actualHex = MurMur3.HashToUInt32(bytes, seedHex);
 
         actualHex.Should().Be(expectedHex);
-        
+
         var offsetBytes = Encoding.UTF8.GetBytes("123" + input + "321");
         var actualOffsetHex = MurMur3.HashToUInt32(offsetBytes, 3, offsetBytes.Length - 6, seedHex);
